+++ test/e2e/samples_test.go
@@ -29,7 +29,7 @@ import (
 )
 
 func TestSamples(t *testing.T) {
-	sampleFiles, err := filepath.Glob("../../config/samples/*/*.yaml")
+	sampleFiles, err := filepath.Glob("../../config/samples/logstash/logstash_stackmonitor.yaml")
 	require.NoError(t, err, "Failed to find samples")
 
 	decoder := helper.NewYAMLDecoder()
